New Enemy Type: Beamers
This formidable enemy type fires a continuous beam for 1.5s. Moves diagonally left or right to fire over as much of the screen as possible. Wait for an opening, then take them out as soon as you can. (Do need to make them visually distinct from the other enemy types)